:bug: Reworked Backend Components to delete the Series

+++ Software Project/1 - Backend - Spring/src/main/java/de/dhbw/ase/whiskey_o_clock/service/SeriesServiceImpl.java
@@ -97,7 +97,18 @@ public class SeriesServiceImpl implements SeriesService {
     */
     @Override
     public void deleteSeries(UUID uuid) {
-        seriesRepository.deleteById(uuid);
+        //Remove Series from Bottles#
+        if(seriesRepository.existsById(uuid)) {
+            Series targetSeries = seriesRepository.getSeriesByUuid(uuid);
+            List<Bottle> bottleList = bottleRepository.getBottlesBySeries(targetSeries);
+            if (null != bottleList){
+                for (Bottle bottle : bottleList) {
+                    bottle.setSeries(null);
+                    bottleRepository.save(bottle);
+                }
+            seriesRepository.deleteById(uuid);
+            }
+        }
     }
     /************************************************************************************************************************************/
